Separation of special classes into new renpycompat module - moved defintion of renpy fake module from "__init__" - defined shorthands to magic module - corrected related imports - adjusted "__all__" exports
