+++ unrpyc.py
@@ -48,91 +48,14 @@ except ImportError:
             pass
 
 import decompiler
-from decompiler import magic, astdump, translate
+import deobfuscate
+from decompiler import astdump, translate
+from decompiler.renpycompat import (pickle_safe_loads, pickle_safe_dumps, pickle_safe_dump,
+                                    pickle_loads)
 
-# these named classes need some special handling for us to be able to reconstruct ren'py ASTs from pickles
-SPECIAL_CLASSES = [set, frozenset]
-
-# ren'py _annoyingly_ enables fix_imports even in ren'py v8,and still defaults to pickle protocol 2.
-# so set/frozenset get mapped to the wrong location (__builtins__ instead of builtins)
-# we don't want to enable that option as we want control over what the pickler is allowed to unpickle
-# so here we define some proxies
-class oldset(set):
-    __module__ = "__builtin__"
-oldset.__name__ = "set"
-SPECIAL_CLASSES.append(oldset)
-
-class oldfrozenset(frozenset):
-    __module__ = "__builtin__"
-oldfrozenset.__name__ = "frozenset"
-SPECIAL_CLASSES.append(oldfrozenset)
-
-@SPECIAL_CLASSES.append
-class PyExpr(magic.FakeStrict, str):
-    __module__ = "renpy.ast"
-    def __new__(cls, s, filename, linenumber, py=None):
-        self = str.__new__(cls, s)
-        self.filename = filename
-        self.linenumber = linenumber
-        self.py = py
-        return self
-
-    def __getnewargs__(self):
-        if self.py is not None:
-            return str(self), self.filename, self.linenumber, self.py
-        else:
-            return str(self), self.filename, self.linenumber
-
-@SPECIAL_CLASSES.append
-class PyCode(magic.FakeStrict):
-    __module__ = "renpy.ast"
-    def __setstate__(self, state):
-        if len(state) == 4:
-            (_, self.source, self.location, self.mode) = state
-            self.py = None
-        else:
-            (_, self.source, self.location, self.mode, self.py) = state
-        self.bytecode = None
-
-@SPECIAL_CLASSES.append
-class Sentinel(magic.FakeStrict, object):
-    __module__ = "renpy.object"
-    def __new__(cls, name):
-        obj = object.__new__(cls)
-        obj.name = name
-        return obj
-
-# These appear in the parsed contents of user statements.
-@SPECIAL_CLASSES.append
-class RevertableList(magic.FakeStrict, list):
-    __module__ = "renpy.revertable"
-    def __new__(cls):
-        return list.__new__(cls)
-
-@SPECIAL_CLASSES.append
-class RevertableDict(magic.FakeStrict, dict):
-    __module__ = "renpy.revertable"
-    def __new__(cls):
-        return dict.__new__(cls)
-
-@SPECIAL_CLASSES.append
-class RevertableSet(magic.FakeStrict, set):
-    __module__ = "renpy.revertable"
-    def __new__(cls):
-        return set.__new__(cls)
-
-    def __setstate__(self, state):
-        if isinstance(state, tuple):
-            self.update(state[0].keys())
-        else:
-            self.update(state)
-
-class_factory = magic.FakeClassFactory(SPECIAL_CLASSES, magic.FakeStrict)
 
 printlock = Lock()
 
-# needs class_factory
-import deobfuscate
 
 # API
 
@@ -156,16 +79,11 @@ def read_ast_from_file(in_file):
         chunks[slot] = raw_contents[start: start + length]
 
     raw_contents = zlib.decompress(chunks[1])
-
-    # import pickletools
-    # with open("huh.txt", "wb") as f:
-    #     pickletools.dis(raw_contents, out=f)
-
-    data, stmts = magic.safe_loads(raw_contents, class_factory, {"_ast", "collections"})
+    _, stmts = pickle_safe_loads(raw_contents)
     return stmts
 
 
-def decompile_rpyc(input_filename, overwrite=False, dump=False, 
+def decompile_rpyc(input_filename, overwrite=False, dump=False,
                    comparable=False, no_pyexpr=False, translator=None,
                    init_offset=False, try_harder=False, sl_custom_names=None):
 
@@ -212,7 +130,8 @@ def extract_translations(input_filename, language):
     translator = translate.Translator(language, True)
     translator.translate_dialogue(ast)
     # we pickle and unpickle this manually because the regular unpickler will choke on it
-    return magic.safe_dumps(translator.dialogue), translator.strings
+    return pickle_safe_dumps(translator.dialogue), translator.strings
+
 
 def parse_sl_custom_names(unparsed_arguments):
     # parse a list of strings in the format
@@ -254,7 +173,8 @@ def worker(t):
         else:
             if args.translation_file is not None:
                 translator = translate.Translator(None)
-                translator.language, translator.dialogue, translator.strings = magic.loads(args.translations, class_factory)
+                translator.language, translator.dialogue, translator.strings = (
+                    pickle_loads(args.translations))
             else:
                 translator = None
             return decompile_rpyc(filename, args.clobber, args.dump, no_pyexpr=args.no_pyexpr,
@@ -391,10 +311,10 @@ def main():
                 bad += 1
                 continue
             good += 1
-            translated_dialogue.update(magic.loads(result[0], class_factory))
+            translated_dialogue.update(pickle_loads(result[0]))
             translated_strings.update(result[1])
         with open(args.write_translation_file, 'wb') as out_file:
-            magic.safe_dump((args.language, translated_dialogue, translated_strings), out_file)
+            pickle_safe_dump((args.language, translated_dialogue, translated_strings), out_file)
 
     else:
         # Check per file if everything went well and report back
